pyproject.toml
[project]
name = "carwash"
version = "0.1.0"
description = "6 semestr Carwash backend app"
authors = [
    {name = "stoicismguy",email = "pyzan55@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
